Gazebo world project
Gazebo is Robot simulation is an essential tool in every roboticist's toolbox. A well-designed simulator makes it possible to rapidly test algorithms, design robots, perform regression testing, and train AI system using realistic scenarios. Gazebo offers the ability to accurately and efficiently simulate populations of robots in complex indoor and outdoor environments. At your fingertips is a robust physics engine, high-quality graphics, and convenient programmatic and graphical interfaces. Best of all, Gazebo is free with a vibrant community. 
Concepts and Classes
Concepts explored in this project:

Gazebo model and world-building
Gazebo plugins
C++
To view this project, you must have Gazebo installed on Linux. Click here for Gazebo download and installation instructions.

With Gazebo installed, clone the repository and navigate to the root directory. You first must build the plugin, so create a build folder and navigate to it:

$ mkdir build && cd build
Then build the plugin:

$ cmake .. && make
And update the GAZEBO_PLUGIN_PATH environment variable:

$ export GAZEBO_PLUGIN_PATH=${GAZEBO_PLUGIN_PATH}:${PWD}
Once the plugin is built, open the world in Gazebo:

$ gazebo ../world/timsWorld
The world should open in Gazebo, and the terminal should display the message:

hello world!